Fixed [Kanji1] code
I say "kanji", but it actually switches to the first character table. That is, the kana table. So kanji is a lie. But I don't care enough to switch the code to [Kana] in every dump and decrement the other kanji codes as appropriate. In other words, it actually displays the proper kana after switching back.